unix/stream: Fix send(2) stall
send(2) on a unix/stream socket can stall indefinitely if the low water
mark of the receiver has been set to a value higher than the available
space, which should not be illegal.

Copy the logic of sosend_generic_locked to not wait if there is enough
space for the send, regardless of the low water mark of the receiver.

This restores functionality that was lost with the new unix/stream
implementation.

Signed-off-by: Ryan Moeller <rmoeller.dev@gmail.com>
Fixes: d15792780760 ("unix: new implementation of unix/stream & unix/seqpacket")