sbin/ifconfig: Accept and display alternative carp(4) key formats
A CARP key was intended to be the 20-byte SHA-1 hash of the passphrase,
but this was never completed on the userland side.  Instead, we have
been silently truncating the passphrase to 19 characters and using that
directly as the key.

Make ifconfig hash the passphrase given with the "keypass" parameter to
adhere to the original design goal and improve the entropy distribution.

SHA-256 is chosen over SHA-1 because the latter is known to have poor
collision avoidance.  The SHA-256 hash is used according to NIST SP
800-56C Rev. 2 Section 4 One-Step KDF (IG D.P ... Without a Counter).

To provide an unhashed key string directly, a new "key" parameter is
added.  Keys that would be truncated are rejected.

A second new alternative parameter "keyhex" accepts the raw key as a
40-character ASCII string representing the 20-byte key in hexadecimal.

The existing behavior of the "pass" parameter is retained for
compatibility, including silent truncation.

Printing the key now uses a heuristic that prints the key as text if its
contents are a string of one or more printable characters followed by a
string of one or more nul bytes for the remainder, otherwise the key is
printed in hexadecimal.

Signed-off-by: Ryan Moeller <rmoeller.dev@gmail.com>